Minor changes and README.md data

+++ src/models/staff.js
@@ -3,8 +3,10 @@ const database = require('../../database/file-database');
 exports.createStaffMember = async (name, salary, currency, on_contract, department_id, sub_department_id) => {
 	//create new staff member and return the staff member data
 	//set on_contact to 0 if it's not provided
-	if (on_contract === undefined) {
+	if (on_contract === undefined || !on_contract) {
 		on_contract = 0;
+	} else {
+		on_contract = 1;
 	}
 
 	const sql = `INSERT INTO staff (name, salary, currency, on_contract, department_id, sub_department_id) VALUES ('${name}', ${salary}, '${currency}', ${on_contract}, ${department_id}, ${sub_department_id})`;
@@ -21,8 +23,13 @@ exports.createStaffMember = async (name, salary, currency, on_contract, departme
 
 exports.deleteStaffMember = async (id) => {
 	//delete staff member
+	if (isNaN(id) || id <= 0) {
+		const error = new Error('Invalid staff member id');
+		error.statusCode = 400;
+		throw error;
+	}
 	const sql = `DELETE FROM staff where id = ${id}`;
-	//delete the record and return true
+	//delete the record and return trues
 	//it can be more sophisticated to check if the member exists before deleting
 	await database.query(sql, 'run');
 	return true;
